Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> At a bare minimum this is going to need a design doc and a bug.
> 
> cuttlefish is supposed to be like other devices, so I'm a bit skeptical about the attempt to add a new partition.

I agree, can you update this CL with a Bug: tag and put the details of why we need this in the bug?

I think it's OK to have the build system generate the metadata partition, as long as it would still work to enable this feature on pixel and flash it with "fastboot flash metadata metadata.img" and not *rely* that it goes through "fastboot erase metadata". Otherwise we are encoding a platform difference which could not be tested in a common way.

If we introduce a bootloader for cuttlefish in the future we can always roll this back.

Patch-set: 1
Reviewer: Gerrit User 1293738 <1293738@85c56323-6fa9-3386-8a01-6480fb634889>