Removing GSI signing
GSI is now configured as a chained partition for signing.
However, this makes emulator cannot enable dynamic partitions because
system's vbmeta is in the logical system partition.

This CL removes the system.img signing and relies on the offline signing
process for GSI. Also, we cannot set AVB_VBMETA_IMAGE_FLAGS_VERIFICATION_DISABLED
for vbmeta.img because it will be used by emulator now. For CTS-on-GSI,
devices can flash its own vbmeta.img with --disable-verification.

Bug: 128434470
Test: checks GSI can be signed by the release signing process independently
      ./build/tools/releasetools/sign_target_files_apks \
        --avb_system_algorithm SHA256_RSA2048 \
        --avb_system_key external/avb/test/data/testkey_rsa2048.pem \
      out/dist/*-target_files-*.zip signed-target_files.zip

Change-Id: I71b0ac407b5676b7cb32db8c9e652e1b4ad07b48