Adding boot-debug.img and ramdisk-debug.img
The two new debugging images will only be built on user build.
Where there is a /force_debuggable file in the ramdisk.
When this file is present, /init will load additional sepolicy
and/or properties from this ramdisk to allow adb root, if the
device is unlocked.

It's intentional to skip signing for boot-debug.img, as it can
only be used if the device is unlocked, where verification error
is allowed.

Those debugging images will be built dy default on user targets.
This can faciliate more testings, which used to run on userdebug images
because of 'adb root', on user build images. This is helpful to narrow
down the delta between what's being tested v.s. what's being shipped.

Bug: 126493225
Test: 'make' on a user build, checks both boot-debug.img and
      ramdisk-debug.img are in $OUT/.
Test: 'make dist' on a user build, checks both images are in out/dist
Test: system/core/mkbootimg/unpack_bootimg.py --boot_img boot-debug.img,
      checks the extracted out/ramdisk is as expected
Test: gunzip -c ../ramdisk-debug.img | cpio -idm, checks it adds
      /force_debuggable from the original ramdisk.img or ramdisk-recovery.img

Change-Id: I30137c3caef91805d9143d404e5e4d06c0fccc30