Adding GSI public keys.
Adds the following line into a device.mk to include GSI keys into the
first-stage ramdisk.

   $(call inherit-product, $(SRC_TARGET_DIR)/product/gsi_keys.mk)

For A/B devices with BOARD_USES_RECOVERY_AS_BOOT set to true, use the
following to put keys into the first stage ramdisk coexisted in /boot.

   $(call inherit-product, $(SRC_TARGET_DIR)/product/gsi_keys_recovery_as_boot.mk)

The GSI keys allow us to boot a GSI with AVB to enable dm-verity.

   https://android.googlesource.com/platform/external/avb/

The keys for R and S GSI is to allow a device boots a new
version of GSI in a Treble-compatible manner.

For more information about GSI, please visit:

    https://source.android.com/setup/build/gsi

Bug: 112293933
Test: make checks $OUT/ramdisk/avb/ and $OUT/recovery/root/first_stage_ramdisk/avb
Change-Id: I8e4a9a8dd607916a3f45a07e0ba14c58def5d8d3