Add more product defaults to mainline.mk
Replicate aosp_blueline's configuration, which is shared with other
pixels, and explicitly repurpose this to be a generic AOSP mainline
device.

Ignore WAPPushManager, which I believe to be obsolete. I will
remove this from blueline.

Clean up associated whitelist entries in gsi for the apps I'm
moving to /product in the same topic.

Bug: 129077057
Test: boot aosp_blueline
Test: browse ringtones
Change-Id: I8424cb23923987d2176369f93b479fc827fbabef