Removed an obsolete TODO.
The intended check was to ensure system image won't be built as a
chained partition when using dynamic partitions. Because such an image
can't be verified by bootloader or kernel.

As it evolves, GSI system image has been built against the intended
check. Because GSI system image will be verified by first stage init in
userspace.

Bug: 80195851
Test: TreeHugger
Change-Id: I45b079799e5d7443e82d73d43e2b63c613f33165