Update patch set 6
Patch Set 6:

> Fundamentally neither this change nor what I've been planning on looking at (PRODUCT_HOST_PACKAGES split just finished at the end of last week) solve the root problem here -- if you don't want these packages to be on system, there need to be no rules to install them onto system. Otherwise the root problem is just being ignored by making everything else behave nicely, and that can be extremely fragile.

You're right, it's fragile. b/129006418 is about getting rid of the build rules, but that can only happen when there's a story for gtests (b/129539670 and b/129534335). I think all of those needs addressing, but in the meantime there's the library absence test in aosp/930987 that (although a bit annoying - b/129466457) will ensure we don't regress due to the fragility. However, as long as this isn't in we can't tighten up the list of exceptions to the absence test, and having the libraries duplicated both takes extra space and can cause ODR bugs.

Patch-set: 6