Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> Why does "module is not installed" imply "shared library dependencies are not setup"?
> 
> Can't the shared library dependencies be fixed?
As it is now, DEPENDENCIES_ON_SHARED_LIBRARIES expect each entry to contain the installed module path, and that doesn't exist for module not installed. Even though the installed module path is not used in this case, I'm hesitant to add entries that has empty field, as that may easily cause confusing bugs down the road. (In fact, if we simply comment out the "ifdef LOCAL_INSTALLED_MODULE" around DEPENDENCIES_ON_SHARED_LIBRARIES in soong_cc_prebuilt.mk, we get a build error. We can change that ifdef to a more complicated condition, but that'd be coupling soong_cc_prebuilt.mk and main.mk more.)

We can add another variable to hold the mapping between modules and their shared library dependencies, but that adds more maintenance burden. My naive assumption is that the purpose of ELF file check is to ensure all dependencies are installed so that we don't get missing library errors at runtime. Assuming that is indeed the only purpose of this check, I think adding more maintenance cost may not be worth it. That said, I don't have the full picture of why we have this check, so I'm open to suggestion on how best to fix this.

Patch-set: 1
Reviewer: Gerrit User 1349653 <1349653@85c56323-6fa9-3386-8a01-6480fb634889>