Update patch set 1
Patch Set 1:

This is only a problem when no-vendor-variant VNDK is enabled, where those VNDK libraries have their vendor variant as uninstallable. I'm trying to enable this flag for the first target now so that this kind of issues will be caught earlier. This is also why you didn't see any issue before.

Patch-set: 1