Unset AB_OTA_UPDATER for mainline.
The previous need to build mainline devices as A/B no longer applies
[1], as we have unbundled the install of update-payload-key.pub.pem from
AB_OTA_UPDATER. Unsetting AB_OTA_UPDATER from BoardConfigMainlineCommon
defers the decision of using A/B OTA to be board-specific.

mainline_arm64 is considered as a specific board to build generic
system-only OTAs in A/B format. So it defines the two variables there.

[1] commit 571fab0610dc51afafe3d33c1ab460b61adeba00

Bug: 130516531
Test: TreeHugger
Test: Compare the built images for GSI and mainline targets. They don't
      have different artifacts due to the change.
Change-Id: I33cc9b6623dca7cc7c2a7e2c65cdbccced5e816e