Deprecate next set of make module types
 BUILD_AUX_EXECUTABLE
 BUILD_AUX_STATIC_LIBRARY

These shouldn't be used anymore, they were an intermediate state to
build some firmware that is no longer being built in Android.

 BUILD_HOST_FUZZ_TEST
 BUILD_HOST_STATIC_TEST_LIBRARY
 BUILD_NATIVE_BENCHMARK
 BUILD_STATIC_TEST_LIBRARY

There are no users of these left.

 BUILD_HOST_STATIC_LIBRARY
 BUILD_HOST_NATIVE_TEST

There are not any common users of this left in AOSP. There are a few
left in device-specific code in internal master, but this just turns it
into a warning.

Bug: 130691889
Bug: 130723134
Bug: 130723227
Bug: 130734558
Test: build_test & check logs
Change-Id: I8a9385040316495f7e7262d045be871e91712920