Using test key to sign boot-debug.img if needed
When chain partition is used for /boot partition, libavb will try to
read AVB footer from there, even if the device is unlocked. However,
this makes device unbootable when boot-debug.img, which is not signed at
all, is used. This CL uses BOARD_AVB_BOOT_KEY_PATH to sign
boot-debug.img if it's defined, which means /boot partition is chained.

The release key of boot.img will be used when invoking
sign_target_files_apks.py offline, during the release signing process.
BOARD_AVB_BOOT_KEY_PATH in the codebase is not a release key.

Bug: 129508966
Test: Enables chain partition for boot.img, `make bootimage_debug` then
      checks `avbtool info_image --image $OUT/boot-debug.img`
Test: `make bootimage_debug-nodeps`

Change-Id: I870cb70c70b7d4d7a30b77bed58cbca6e007d5e3