Update patch set 2
Patch Set 2: Code-Review+1

> They believe these aren't the right tools to use for debugging early startup processes. 

Ok, then I am fine with removing this from /system.

> Even if they could be useful with the bootstrap Bionic, they ought to reside in lib*/bootstrap rather than one level up, right?

No, having the lib directly under /system/lib is fine. Having the bootstrap libc.so under ./bootstrap is just because we would like to use the legacy path /system/lib/libc.so as a symlink to the real libc.so in the APEX... for compatibility reason. For other libraries where such compatibility issue doesn't exist, /system/lib/libFoo.so could be the one for the non-updatable part of the platform.

Patch-set: 2
Reviewer: Gerrit User 1132673 <1132673@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1