Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> I'm not 100% sure about libFuzzer.a being an implicit dependency through -fsanitize=fuzzer.
> Is it added by the driver even with -nostdlib?

It pulls in libclang_rt.fuzzer-${arch}.a. We depend on lots of libc things in fuzzer, is it even supported to try and "-fsanitize=fuzzer -nostdlib"?

Looks like we need to statically link here against libFuzzer.a, as the toolchain doesn't pull in libclang_rt.fuzzer-${arch}.a...

Patch-set: 3
Reviewer: Gerrit User 1453063 <1453063@85c56323-6fa9-3386-8a01-6480fb634889>