Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Where do you expect TARGET_SKIP_CURRENT_VNDK to be set?
> How does this work with mainline?

system/lib/vndk[-sp]-{ver} is for the specific vendor versions with {ver}.
However, current system images are always include the VNDK version for current platform version.
It is waste of space if we don't have plan to install the system image on the same vendor version. We want to provide a method to skip installing the current VNDK version.

Patch-set: 2
Reviewer: Gerrit User 1146974 <1146974@85c56323-6fa9-3386-8a01-6480fb634889>