make: Ensure recoveryimage build depends on componants needed
Several parts of the recovery image, such as dtbo images, are passed
to the recovery image generation command without any checks to see if
these parts are even present. Add them to RECOVERYIMAGE_EXTRA_DEPS to
have the build system ensure their presence before making the image.

Also, several variables are expected to be set to true/false rather than
actual paths or arguments, so make sure we only check to see if those are
set to true rather than if they're set to anything.

Test: m -j recoveryimage with BOARD_INCLUDE_RECOVERY_DTBO := true
      and BOARD_PREBUILT_DTBOIMAGE set to non-existent file. Build stops
      as expected.
Test: m -j recoveryimage with BOARD_INCLUDE_RECOVERY_DTBO := false
      recovery.img contains no DTBO.
Change-Id: I50cf5f97773478b94140a3a6b31712a68513eba5