Generate block based OTA for product partitions
Currently only the system and vendor partitions are updated in the
generic block based OTA generation script. Since the product partition
is quite large and consist most of APK files, it would be beneficial to
update it similar to system.

This also tries to resolve the issue where the device script adds
arbitrary progress resulting in a OOB progress bar during the update.

Bug: 132683080
Bug: 139556234
Test: Run unit tests, generate full and incremental OTA with product partitions.
Change-Id: I13478cf9bd32137c6729b8c9cb102080147093f2