Allow sanitizers with forced static binaries in Makefiles.
Originally, sanitizers were incompatible with forced static binaries,
due to the clang driver explicitly trying to link the sanitizer's shared
library during link time. This fails as as you can't statically link a
dynamic library. The fix is to explicitly set '-static-libsan' in the
linker flags.

This patch addresses a TODO as well, as it allows diagnostic sanitizers
to be used with static binaries.

Bug: 121042685
Test: SANITIZE_TARGET='fuzzer' m
Change-Id: I01dd1a7d242378ac2a340da721955c00c9259cb5