Support building recovery for non-sparse images.
There is a check that enforces that the recovery image must fit in the
space between the end of the system filesystem and the end of the system
partition, which is impossible when the system partition is not sparse.

This change disables that check when BOARD_USES_FULL_RECOVERY_IMAGE is
true, which implies that the complete recovery image is already stored
inside the system filesystem, making it redundant to store it anywhere
else.

Test: Add BOARD_USES_FULL_RECOVERY_IMAGE, remove TARGET_NO_RECOVERY in cuttlefish
Bug: 120041970
Change-Id: Id1448a3e4ef57323e5c762a3a319e2dce988a888