Update patch set 1
Patch Set 1:

We tend to push back on random flags like this, they make builds harder to test, and cause more support issues when they are used incorrectly.  What is the impact of this flag?  How much faster does it make the build?  I assume if it was safe to disable AVB it would always be disabled, so what is going to break if someone turns on this flag without understanding it?

Patch-set: 1
Reviewer: Gerrit User 1002751 <1002751@85c56323-6fa9-3386-8a01-6480fb634889>