Move extra licenses to vendor from system NOTICE file
If building with with a vendor image, ensure that the licenses from
extra partitions are included in the vendor NOTICE file and not the
system NOTICE file since they could come from disperate source trees.
This change does not affect existing NOTICE files outside of the system
partition. (e.g. product and product_services.

Bug: 13467960
Test: Built a vendor image enabled target and ensured the notices were in the
proper file
Change-Id: Ib17c7d7f84805734de09e686a4835b9b680b4192