Fix symlink dependencies.
The order-only dependency is not sufficient. If the target of the
symlink is rebuild, anything depending on it would not be rebuild.

This fixes incremental builds.  Clean build will build everything.

Technically, the comment is correct, however the timestamp change
would be noticed on the next build, requiring building in a loop.

Test: Fixes ART tests for incremental builds.
Change-Id: Ic3b93ee98fda2dc707435d35053b437f81efb7b0