Allow empty framework misc_info entries in merge_builds
If a key listed in the framework_misc_info_keys file has no corresponding
value in the framework misc_info.txt, let that absence be reflected in
the merged misc_info, instead of falling back to the vendor value. This
is necessary to avoid the use of the vendor avb_system_key_path value
when regenerating vbmeta.

Test: manual
Bug: 137184321
Change-Id: I7eafe6e1de50548d97841f93bd97cd8d8d3c81ff