check_elf_file.mk: checks correspond to arguments
Before this, it's really expensive to turn on PRODUCT_CHECK_ELF_FILES
for every device. For instance, especially on old devices, it's unlikely
we'll ever really be able to add "LOCAL_SHARED_LIBRARIES" to every
prebuilt in order to declare the right dependencies.

However, some checks, like 'soname matching' can be turned on globally.

Additionally, if we add an option like 'no_prebuilt', we can also turn
this on globally (b/140559974). This is because per-device bringup
doesn't really give us an incremental bringup (although shipping api
level >= level might work for options like this).

Bug: 139525855
Test: m check-elf-files
Change-Id: I7926bbca6744ee401de6e2d3475ad9d035ac227b