Update patch set 2
Patch Set 2:

I checked internally and there's no merging from this branch to others.  It's meant to be a snapshot of android-10 and not a working branch.

I did file a bug to try to get this change included in future releases though I haven't heard anything back from the release team about that matter.

Especially since this is a build change, and won't impact the artifacts created in any of the images, I think your best solution is going to be to cherry-pick this into your android-10 builds.

Patch-set: 2
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>