Update patch set 5
Patch Set 5: Presubmit-Ready+1

For those of you playing along at home, the presubmits are an incremental build, and CleanSpec is a special script that when lines are added, they get run. We need to clean up every *fuzz* folder in the system from these changes because the fuzz targets have moved into a subdirectory. This causes a huge kerfuffle, and I currently have to whack-a-mole every single one of {product out, product out (symbols), host out, soong host out (i have no idea what this is)}.

I'm really hoping that this is now the last thing that I have to add to that list, because there is NO WAY for me to test this in my local dir. Each one is a 2hr presubmit round trip. (╯°□°)╯︵ ┻━┻

Patch-set: 5
Reviewer: Gerrit User 1453063 <1453063@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=+1