Create change
Uploaded patch set 1.

Patch-set: 1
Change-id: Ia4045bd67ffb3d899efa8d20dab4c4299b87ee5f
Subject: This change is part of a topic that moves the recovery resources from the system partition to the vendor partition, if it exists, or the vendor directory on the system partition otherwise. The recovery resources are moving from the system image to the vendor partition so that a single system image may be used with either an A/B or a non-A/B vendor image. The topic removes a delta in the system image that prevented such reuse in the past.
Branch: refs/heads/master
Status: new
Topic: 
Commit: 927504754bd03b981a7d1c627950d38161e83d04
Tag: autogenerated:gerrit:newPatchSet
Groups: 927504754bd03b981a7d1c627950d38161e83d04
Private: false
Work-in-progress: false