Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> Do we want to hold off on this to make it easier to merge the already existing changes to warn.py from the ChromeOS side? This refactoring/addition seems like it might make that harder to do.

It looks like the merge is going to take a fair bit of re-working on our side anyway (we're currently depending on some Google3 stuff that apparently Android can't use), so I don't think holding off on these changes for the merge is important.  I am trying to keep a list of all the changes Chih-hung is making, so I can apply them as I work on doing the merge.

Patch-set: 2