make: Make dtb.img depend on the DTB directory itself
We shouldn't be depending on .dtb files for this, because if the directory
doesn't exist for some reason, the DTBs just won't get added to the image.
Instead depend on the directory itself.

Also, the current cat command generates static rules with a list of DTBs
that it finds when build rules are generated. This isn't good, especially
when we build our own DTBs, so rewrite it to find DTBs at "command running
time" instead.

Test: Remove device/google/bonito-kernel and attempt to build;
      Build fails due to missing device/google/bonito-kernel
Test: Remove device/google/bonito-kernel/sdm670.dtb and build;
      Empty dtb.img created
Test: Build b1c1 dtb.img before and after change and verify it's
      the same image
Change-Id: Ia9336e9767999bc792aa977a73ae9739a4bc2c37