Update patch set 3
Patch Set 3:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > so, this may not be the best place, but what do you think about this? or could you suggest another place?
> 
> Would we want a separate place within Google's internal master for changes made on projects in internal master?  Presumably we wouldn't want to have those changes published in AOSP before the changes get pushed to AOSP.
> 
> Also, I know the build team has their own file for build system related changed.  I don't want to break what they're doing, since they've been on top of this for a while.

This is intentionally separate per-file, so changes in different branches could be added without any risk for merge conflicts.

Changes.md is separate here, but it is prone to this merge conflict file and it is specific to build/make changes. I'm more interested in creating a location that is for any part of android, but which is also centralized for convenience.

I think we could consider doing something distributed as well, like CleanSpec.mk, but say 'FixSpec.mk'. However, it has proven quite painful to have many scripts like this. Perhaps we could reserve a directory like this for large changes spanning all branches of android and make it distributed as/if there becomes a need.

Patch-set: 3